Created julia.shader and ColorRect.gd
julia.shader is a fragment shader that draws Julia Sets. ColorRect.gd is a script that changes the value of c at a certain speed to create animations using julia.shader. For now you can only change this via coding and compiling but I'm planing on adding a GUI to change every variable. Also I added some controls and changed a few things.